refactor Program.cs
